requirements.txt
dask==2023.10.1
duckdb==0.10.0
pandas==1.5.3
psutil==5.9.0
PyYAML==6.0.1
Requests==2.31.0
tqdm==4.64.1
